+++ packages/stylelint-config-terra/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * Breaking
   * Drop support for node versions below version 10.
-  * Require stylelint version 13
+  * Require stylelint version 13.
+  * Use jest for tests.
 
 ## 3.7.0 - (July 28, 2020)